Added Better Support for Other Openstack Clouds
This commit now adds the ability for a user to leverage multiple openstack
clouds effortlessly.  I have added an "os-hp-auth" plugin to the command line
options which will allow an HP cloud consumer to use Turbolift to command
Swift "Object Store" as well as the CDN.

I have also improved the auth plugins to be more intutive as well as the
auth_utils.

Http utils is no longer tied to Rackspace public cloud when using
a url which has no prefix.