Fixes regression in Version 2.0.5
This commit fixes an issue with upload when using the --delete-remote
flag.  If this flag is used the entire contents of the container will
be marked for deletion as the index list is no longer a percistance list.

This fixes that issue by reindexing the local file system prior to
determining what will be marked for deletion in the remote container.