This commit enforces unicode on a local index
By default python 2 uses byte encoding and will attempt to
compare byte encoded strings with unicode strings which will
equal True due to the internal workings of python. With that
said on some file systems it is likely that the program will
encounter non-ascii compatible stings. This change ensures
that all indexed files are encoded as unicode from the very
beginning.