Further work on unicode indexing
It seems that when a file system is NTFS and the base system
encoding is "utf-8" a potential issue happens when files were created
on the NTFS filesystem and encoded using some "windows" encoding type,
specifically when attempting to encode / decode byte strings. To solve
this I changed everything to be encoded as unicode though this alone
can cause other fun issues. To itereate on a solution here I've gone
and broke out the indexer into parts that will attempt to perform the
decoding as either UTF-8 and then reencode as unicode for use everywhere
else.