Fixed quote issue and typo
The quoter function was not able to deal with a unicode encoded
unicode string when running on Python2.7. When something like this
was passed it would result in a KeyError. To fix this a "try/Except"
was added to the system such that it allows the origial value to be
passed back from the method in such a case. This was done because
we can safely assume that a unicode encoded unicode string is already
in a state that has been quoted for use both remotely and locally.

Closes-Bug: #94

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>